requirements.txt
acme==0.31.0
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
Babel==2.8.0
blinker==1.4
certbot==0.31.0
certbot-nginx==0.31.0
certifi==2018.1.18
chardet==3.0.4
click==6.7
cloud-init==18.3
colorama==0.3.7
command-not-found==0.3
ConfigArgParse==0.11.0
configobj==5.0.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.1.4
defusedxml==0.6.0
diff-match-patch==20181111
distro-info==0.18ubuntu0.18.04.1
Django==2.2.2
django-braces==1.14.0
django-ckeditor==5.7.1
django-colorfield==0.2.1
django-cors-headers==2.5.2
django-crontab==0.7.1
django-environ==0.4.5
django-filter==2.1.0
django-guardian==1.5.0
django-import-export==2.0.2
django-jet==1.0.8
django-js-asset==1.2.2
django-oauth-toolkit==1.2.0
django-phonenumber-field==4.0.0
django-rest-framework-social-oauth2==1.1.0
django-rest-swagger==2.2.0
django-simple-history==2.7.3
djangorestframework==3.9.2
djangorestframework-guardian==0.1.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.1.0
environ==1.0
et-xmlfile==1.0.1
future==0.15.2
gunicorn==19.9.0
httplib2==0.9.2
hyperlink==17.3.1
idna==2.6
incremental==16.10.1
itypes==1.1.0
jdcal==1.4.1
jet==1.0
Jinja2==2.10
josepy==1.1.0
jsonfield==2.0.2
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
MarkupPy==1.14
MarkupSafe==1.0
mock==2.0.0
ndg-httpsclient==0.4.4
netifaces==0.10.4
oauthlib==2.0.6
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.3
PAM==0.4.2
parsedatetime==2.4
passlib==1.7.1
pbr==3.1.1
phonenumbers==8.12.0
Pillow==5.4.1
pip==9.0.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycrypto==2.6.1
pygobject==3.26.1
PyJWT==1.5.3
PyMySQL==0.9.3
pyOpenSSL==17.5.0
pyparsing==2.2.0
pyRFC3339==1.0
pyserial==3.4
python-apt==1.6.5+ubuntu0.2
python-debian==0.1.32
python3-openid==3.1.0
pytz==2018.9
pyxdg==0.25
PyYAML==3.12
requests==2.18.4
requests-oauthlib==1.3.0
requests-toolbelt==0.8.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
service-identity==16.0.0
setuptools==39.0.1
simplejson==3.16.0
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sorl-thumbnail==12.5.0
sqlparse==0.3.1
ssh-import-id==5.7
systemd-python==234
tablib==1.1.0
Twisted==17.9.0
ufw==0.35
unattended-upgrades==0.1
uritemplate==3.0.0
urllib3==1.22
wheel==0.30.0
xlrd==1.2.0
xlwt==1.3.0
zope.component==4.3.0
zope.event==4.2.0
zope.hookable==4.0.4
zope.interface==4.3.2